+++ example/hadoop/src/main/java/org/apache/iotdb/hadoop/tsfile/TsFileHelper.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.iotdb.tsfile.hadoop;
+package org.apache.iotdb.hadoop.tsfile;
 
 import org.apache.iotdb.tsfile.file.metadata.enums.TSDataType;
 import org.apache.iotdb.tsfile.file.metadata.enums.TSEncoding;
@@ -32,10 +32,9 @@ import java.io.File;
 import java.io.FileNotFoundException;
 import java.io.IOException;
 
-public class TsFileTestHelper {
+public class TsFileHelper {
 
-
-  private static final Logger logger = LoggerFactory.getLogger(TsFileTestHelper.class);
+  private static final Logger logger = LoggerFactory.getLogger(TsFileHelper.class);
 
   public static boolean deleteTsFile(String filePath) {
     File file = new File(filePath);
@@ -59,9 +58,14 @@ public class TsFileTestHelper {
       int sensorNum = 10;
 
       // add measurements into file schema (all with INT64 data type)
-      for (int i = 0; i < sensorNum; i++) {
+      for (int i = 0; i < 2; i++) {
         schema.registerMeasurement(
                 new MeasurementSchema("sensor_" + (i + 1), TSDataType.INT64, TSEncoding.TS_2DIFF));
+      }
+
+      for (int i = 2; i < sensorNum; i++) {
+        schema.registerMeasurement(
+                new MeasurementSchema("sensor_" + (i + 1), TSDataType.DOUBLE, TSEncoding.TS_2DIFF));
       }
 
       // add measurements into TSFileWriter
@@ -75,13 +79,18 @@ public class TsFileTestHelper {
 
       long timestamp = 1;
       long value = 1000000L;
+      double doubleValue = 1.1;
 
-      for (int r = 0; r < rowNum; r++, value++) {
+      for (int r = 0; r < rowNum; r++, value++, doubleValue = doubleValue + 0.1) {
         int row = rowBatch.batchSize++;
         timestamps[row] = timestamp++;
-        for (int i = 0; i < sensorNum; i++) {
+        for (int i = 0; i < 2; i++) {
           long[] sensor = (long[]) values[i];
           sensor[row] = value;
+        }
+        for (int i = 2; i < sensorNum; i++) {
+          double[] sensor = (double[]) values[i];
+          sensor[row] = doubleValue;
         }
         // write RowBatch to TsFile
         if (rowBatch.batchSize == rowBatch.getMaxBatchSize()) {
