+++ hadoop/src/main/java/org/apache/iotdb/hadoop/tsfile/TSFHadoopException.java
@@ -16,7 +16,7 @@
  * specific language governing permissions and limitations
  * under the License.
  */
-package org.apache.iotdb.tsfile.hadoop;
+package org.apache.iotdb.hadoop.tsfile;
 
 public class TSFHadoopException extends Exception {
